refactor(renderer): give the sidebar its types apart from its view
Nine of the eleven modules that imported `Sidebar.tsx` wanted a type, not the
component: the drag engine, the filtering, the scope, the agent-state
projection and all four stores name `SidebarProps`, `SidebarAgentState` or
`ResolvedPinnedItem` and render nothing. The entry component had become the
feature's type module by accident, so every piece of pure logic pulled in the
whole view to borrow a name.

`sidebar-types.ts` holds the three types; `Sidebar.tsx` imports `SidebarProps`
like everyone else and is now nine lines of imports and the markup. The two
remaining edges to it - `WorkspaceSidebar` and the test - are the ones that
actually render a sidebar.

Only visible once both halves were in one directory, which is what the move was
for. No behaviour change: 37 tests pass unchanged.
