+++ src/renderer/src/features/sidebar/sidebar-scope.ts
@@ -10,7 +10,7 @@
 import { createContext, createEffect, onCleanup, useContext } from "solid-js";
 import { createScrollFades } from "../../components/createScrollFades";
 import { createSidebarDragEngine } from "./createSidebarDragEngine";
-import type { SidebarProps } from "./Sidebar";
+import type { SidebarProps } from "./sidebar-types";
 import { createSidebarAnnouncementStore } from "./stores/announcement-store";
 import { createSidebarDataStore } from "./stores/data-store";
 import { createSidebarDragStateStore } from "./stores/drag-state-store";
