+++ src/renderer/src/features/sidebar/sidebar-types.ts
@@ -1,0 +1,61 @@
+import type {
+  AvatarHue,
+  DirectThreadSummary,
+  SidebarLayoutAction,
+  SidebarLayoutSnapshot,
+  TeamPresenceMember,
+} from "@openbot/contracts/ipc";
+import type { AgentProfile } from "../../data";
+import type { SidebarPinnedItem } from "./sidebar-pins";
+
+/**
+ * What the sidebar is, as data. These live apart from `Sidebar.tsx` because
+ * nine of this feature's modules - the drag engine, the filtering, the scope
+ * and every store - name one of these types and none of them render anything.
+ * Reading them from the entry component made the pure logic depend on the whole
+ * view to borrow a name.
+ */
+export interface SidebarProps {
+  serverName: string;
+  onOpenServerSettings: (trigger: HTMLElement) => void;
+  agents: AgentProfile[];
+  activeAgentId: string;
+  showPeople?: boolean;
+  people: TeamPresenceMember[];
+  directThreads: DirectThreadSummary[];
+  activeDirectMemberId: string | null;
+  agentStates: Record<string, SidebarAgentState>;
+  layout: SidebarLayoutSnapshot;
+  layoutMutable?: boolean;
+  collapsedSectionIds: string[];
+  onMutateLayout: (action: SidebarLayoutAction) => Promise<void>;
+  onToggleSection: (sectionId: string) => void;
+  pinnedItems: SidebarPinnedItem[];
+  peopleOrder: string[];
+  onPin: (item: SidebarPinnedItem) => void;
+  onUnpin: (item: SidebarPinnedItem) => void;
+  onReorderPinned: (items: SidebarPinnedItem[]) => void;
+  onReorderPeople: (memberIds: string[]) => void;
+  onSelectAgent: (agentId: string) => void;
+  onSelectPerson: (memberId: string) => void;
+  onPreloadDirectConversation?: () => void;
+  onCreateAgent: () => void;
+  onEditAgent: (agentId: string) => void;
+  duplicateSupported?: boolean;
+  duplicatingAgentIds?: ReadonlySet<string>;
+  onDuplicateAgent?: (agentId: string) => Promise<void>;
+  onDeleteAgent: (agentId: string) => Promise<void>;
+  compact: boolean;
+  onExpand: () => void;
+  onOpenMarketplace: () => void;
+  emptyAction?: {
+    label: string;
+    avatarSeed: string;
+    avatarHue: AvatarHue | null;
+    onSelect: () => void;
+  };
+}
+
+export type SidebarAgentState = { kind: "working" } | { kind: "responded" } | { kind: "unread"; count: number };
+
+export type ResolvedPinnedItem = { ref: SidebarPinnedItem; agent: AgentProfile };
