+++ src/renderer/src/features/sidebar/stores/data-store.ts
@@ -8,9 +8,9 @@ import { SIDEBAR_PEOPLE_SECTION_ID, SIDEBAR_UNASSIGNED_SECTION_ID } from "@openb
 import { createMemo } from "solid-js";
 import { teamMemberName } from "../../../components/TeamPersonAvatar";
 import type { AgentProfile } from "../../../data";
-import type { ResolvedPinnedItem, SidebarProps } from "../Sidebar";
 import { agentMatchesQuery, personMatchesQuery } from "../sidebar-filtering";
 import { sidebarPinnedItemKey } from "../sidebar-pins";
+import type { ResolvedPinnedItem, SidebarProps } from "../sidebar-types";
 
 export function createSidebarDataStore(deps: { normalizedQuery: () => string; props: SidebarProps }) {
   const { normalizedQuery, props } = deps;
